+++ CHANGELOG.md
@@ -36,6 +36,9 @@ Main (unreleased)
   - `loki.write` receives log entries from other `loki` components and sends
     them over to a Loki instance. (@tpaschalis)
 
+  - `loki.relabel` receives log entries from other `loki` components and
+    rewrites their label set. (@tpaschalis)
+
 ### Enhancements
 
 - Update OpenTelemetry Collector dependency to v0.63.1. (@tpaschalis)